Add pipe functionality to 'ision' filter
This commit adds the possibility of invoking external commands to view
or process the output of internal commands using pipes.

Other important change was to add the parameter '-w 200' when invoking
rlwrap. This change was necessary because the default value of 40 was
causing some output to be interpreted wrongly (e.g. when using the
command 'extension list |less' with many extensions in the system).